refactor: remove OrganizationAvatar, People, and UserAvatar components; integrate avatar editing directly into Organization and User settings feat: enhance organization settings with member role management and invitation functionality fix: update login component to safely access oauthAvailability properties
