+++ api/src/kubernetes/templates/application/application.yml
@@ -10,6 +10,7 @@ metadata:
         longlink.io/component: migration
 spec:
     backoffLimit: 0
+    ttlSecondsAfterFinished: 86400
     template:
         metadata:
             labels:
